+++ src/main/java/io/lunamc/protocol/packet/handshake/HandshakeServerboundV47.java
@@ -17,16 +17,7 @@
 package io.lunamc.protocol.packet.handshake;
 
 import io.lunamc.protocol.packet.Packet;
-import io.lunamc.protocol.packet.ProtocolBound;
-import io.lunamc.protocol.packet.PacketMeta;
-import io.lunamc.protocol.packet.ProtocolState;
 
-@PacketMeta(
-        packetId = 0x00,
-        minProtocolVersion = 47,
-        protocolBound = ProtocolBound.SERVERBOUND,
-        protocolState = ProtocolState.HANDSHAKE
-)
 public interface HandshakeServerboundV47 extends Packet {
 
     int getProtocolVersion();
@@ -44,4 +35,9 @@ public interface HandshakeServerboundV47 extends Packet {
     NextProtocolState getNextState();
 
     void setNextState(NextProtocolState nextState);
+
+    @Override
+    default int getPacketId() {
+        return 0x00;
+    }
 }