+++ src/main/java/io/lunamc/protocol/packet/login/SetCompressionClientboundV47.java
@@ -17,19 +17,15 @@
 package io.lunamc.protocol.packet.login;
 
 import io.lunamc.protocol.packet.Packet;
-import io.lunamc.protocol.packet.PacketMeta;
-import io.lunamc.protocol.packet.ProtocolBound;
-import io.lunamc.protocol.packet.ProtocolState;
 
-@PacketMeta(
-        packetId = 0x03,
-        minProtocolVersion = 47,
-        protocolBound = ProtocolBound.CLIENTBOUND,
-        protocolState = ProtocolState.LOGIN
-)
 public interface SetCompressionClientboundV47 extends Packet {
 
     int getThreshold();
 
     void setThreshold(int threshold);
+
+    @Override
+    default int getPacketId() {
+        return 0x03;
+    }
 }